RFC: support bridge local clean session
This is a draft of separating bridge clean session settings between the
local and remote endpoints.  Some broker implmentations refuse to allow
non-clean sessions, as they don't support storing messages to be sent to
the connecting broker.  However, this doesn't meant that the local
broker can't be queueing messages to send _out_ to the remote broker.

This PR adds a new bridge connection setting, local_clean_session, that
allows controlling this split.

INCOMPLETE:  the handling of the defaults is incomplete at this stage.
What is needed is for local clean session to use the value of clean
session, unless it's explicitly set, but I can't find any existing
config examples that have this three way state for booleans.

If this concept is acceptable (that you should be able to split these
clean session settings) I can work on that further.

Signed-off-by: Karl Palsson <karlp@etactica.com>